9.108.1-assn-jpa-customer-operations : refactor deleteCustomerByid service definition and adjust the related implementation when the Customer attempted to be deleted does no exist

+++ src/main/java/guru/springframework/spring6restmvc/services/CustomerServiceJPA.java
@@ -71,8 +71,12 @@ public class CustomerServiceJPA implements CustomerService {
     }
 
     @Override
-    public void deleteCustomerByid(UUID id) {
-        customerRepository.deleteById(id);
+    public Boolean deleteCustomerByid(UUID id) {
+        if(customerRepository.existsById(id)) {
+            customerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
